Split web_compilers_test.dart. Update "integration" test tags.
