+++ build_runner/mono_pkg.yaml
@@ -9,7 +9,7 @@ stages:
   - format
   - analyze: --fatal-infos .
 - test:
-  - test: -x integration1 -x integration2 -x integration3 -x integration4 --test-randomize-ordering-seed=random
+  - test: -x integration1 -x integration2 -x integration3 -x integration4 -x integration5 -x integration6 --test-randomize-ordering-seed=random
   - test: -t integration1 --test-randomize-ordering-seed=random
     os:
     - linux
@@ -26,6 +26,16 @@ stages:
     - macos
     - windows
   - test: -t integration4 --test-randomize-ordering-seed=random
+    os:
+    - linux
+    - macos
+    - windows
+  - test: -t integration5 --test-randomize-ordering-seed=random
+    os:
+    - linux
+    - macos
+    - windows
+  - test: -t integration6 --test-randomize-ordering-seed=random
     os:
     - linux
     - macos
