+++ build_runner/test/integration_tests/build_command_resolve_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Tags(['integration4'])
+@Tags(['integration3'])
 library;
 
 import 'package:test/test.dart';
