+++ build_runner/test/integration_tests/build_command_vm_arg_test.dart
@@ -2,6 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Tags(['integration3'])
+library;
+
 import 'package:test/test.dart';
 
 import '../common/common.dart';
