+++ build_runner/test/integration_tests/daemon_command_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Tags(['integration2'])
+@Tags(['integration4'])
 library;
 
 import 'dart:io';
