+++ build_runner/test/integration_tests/test_command_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@Tags(['integration3'])
+@Tags(['integration5'])
 library;
 
 import 'package:io/io.dart';
